rename from RemoveImageCommant to RemoveBrokenMediaObjectsCommand, divide execute method to many small methods ticket: https://app.asana.com/0/25002076043610/29743393245510
